refactor: back delayFn with node:timers/promises
Replace the hand-rolled `new Promise((resolve) => setTimeout(resolve, ms))`
wrapper in the container with a native promisified timer, and move the
`DelayFn` type out of the Spotify auth service into `src/utils/delay.ts`.

`SpotifyService` was importing a generic sleep type from the auth module,
which coupled the two for no reason. The namespace import is deliberate:
`node:test` mock timers cannot intercept a destructured
`import { setTimeout }`, and `scheduler.wait` is not interceptable either.
