+++ src/services/music-providers/spotify.service.ts
@@ -1,7 +1,8 @@
 import { BaseMusicService } from './base-music.service.js';
 import { LogService } from '../log.service.js';
 import { Playlist, Track } from '../../entities.js';
-import { DelayFn, SpotifyAuthService } from './spotify-auth.service.js';
+import { SpotifyAuthService } from './spotify-auth.service.js';
+import { DelayFn } from '../../utils/delay.js';
 import {
     classifyApiStatus,
     parseRetryAfter,
