+++ src/utils.ts
@@ -1,3 +1,4 @@
 export * from './utils/cleanup.js';
+export * from './utils/delay.js';
 export * from './utils/url.js';
 export * from './utils/array.js';
